requirements.txt
absl-py==1.4.0
astunparse==1.6.3
Babel==2.12.1
bcrypt==4.0.1
cachetools==5.3.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cv==1.0.0
dlib @ file:///C:/Users/20307975/Downloads/dlib-19.22.99-cp310-cp310-win_amd64.whl
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.2.3
Flask-Bcrypt==1.0.1
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
idna==2.10
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.0.1
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.0.4
mysql-connector-python==8.0.32
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.1
Pillow==9.4.0
protobuf==3.20.3
psycopg2-binary==2.9.5
py-bcrypt==0.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2023.3
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
Theano==1.0.5
tk==0.1.0
tkcalendar==1.6.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.6
Werkzeug==2.2.3
wrapt==1.14.1
